+++ src/renderer/components/catalog/Catalog/Catalog.stories.tsx
@@ -3,10 +3,8 @@ import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
 import { Meta } from '@storybook/react';
 import { Catalog } from './Catalog';
-import type {
-	Catalog as CatalogType,
-	CatalogEntry as CatalogEntryType,
-} from '../../../../main/catalog/types';
+import type { Catalog as CatalogType } from '../../../../main/catalog/types';
+import { mockCatalogEntry } from '../mockCatalogEntry';
 
 const meta: Meta = {
 	title: 'Catalog',
@@ -15,56 +13,10 @@ const meta: Meta = {
 
 export default meta;
 
-const completeEntry: CatalogEntryType = {
-	db_id: 'jtcores',
-	gameName: 'Street Fighter Alpha 2 (Euro 960229)',
-	categories: ['Fighting'],
-	titleScreenshotUrl:
-		'https://raw.githubusercontent.com/city41/AMMiSTer/main/screenshots/titles/sfa2.png',
-	gameplayScreenshotUrl:
-		'https://raw.githubusercontent.com/city41/AMMiSTer/main/screenshots/snap/sfa2.png',
-	mameVersion: '0229',
-	manufacturer: ['Capcom'],
-	orientation: 'horizontal',
-	yearReleased: 1996,
-	files: {
-		mra: {
-			type: 'mra',
-			db_id: 'jtcores',
-			fileName: 'Street Fighter Alpha 2 (Euro 960229).mra',
-			md5: 'dummy hash',
-			relFilePath: '_Arcade/Street Fighter 2 (Euro 960229).mra',
-		},
-		rbf: {
-			type: 'rbf',
-			db_id: 'jtcores',
-			fileName: 'jtcps2_20220819.rbf',
-			md5: 'dummy hash',
-			relFilePath: '_Arcade/cores/jtcps2_20220819.rbf',
-		},
-		roms: [
-			{
-				type: 'rom',
-				db_id: 'jtcores',
-				fileName: 'sfa2.zip',
-				md5: 'dummy hash',
-				relFilePath: 'games/mame/sfa2.zip',
-			},
-			{
-				type: 'rom',
-				db_id: 'jtcores',
-				fileName: 'qsound.zip',
-				md5: 'dummy hash',
-				relFilePath: 'games/mame/qsound.zip',
-			},
-		],
-	},
-};
-
 // @ts-expect-error
 const catalog: CatalogType = {
 	updatedAt: Date.now(),
-	jtcores: [completeEntry, completeEntry],
+	jtcores: [mockCatalogEntry, mockCatalogEntry],
 };
 
 export const Basic = () => {